% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{add_mgmt_unit}
\alias{add_mgmt_unit}
\title{Add Management Unit}
\usage{
add_mgmt_unit(sample_frame)
}
\arguments{
\item{sample_frame}{Sampling_Frame name. For all sampling frames, use: Sampling_Frame = "All"}
}
\value{
A tibble copy of "Events_extra_xy" with columns 'Zone' and 'Mgmt' added for each event.
Saved to R folder for now.
}
\description{
Adds Management Unit and Zone Categorical Variables to Events Table and
}
\details{
Saves a new Table (tibble) with management unit for use in other functions.
Management unit is added to the table by performing a spatial join
using both a specific AGOL management layer (hard coded)
and the X,Y plot center coordinates found in the Events table.
}
