% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{add_stats}
\alias{add_stats}
\title{Add stats to vegetation data}
\usage{
add_stats(.data, ...)
}
\arguments{
\item{.data}{dataset with at least 1 numeric parameter}

\item{...}{grouping variables}
}
\value{
datatable and numeric parameter with stats
}
\description{
Add stats to vegetation data
}
\examples{
\dontrun{
Cover <- summarize_understory(plant_grouping = "Nativity")
Cover_Stats <- add_stats(.data = Cover, Unit_Code, Sampling_Frame, Cycle, Stratum, Nativity)

}
}
