% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{add_table_caption}
\alias{add_table_caption}
\title{Create a numbered caption for a DT datatable.
Call this right before or right after the call to DT (depending whether you want the caption on top or bottom)}
\usage{
add_table_caption(caption)
}
\arguments{
\item{caption}{The caption text}
}
\description{
Create a numbered caption for a DT datatable.
Call this right before or right after the call to DT (depending whether you want the caption on top or bottom)
}
