% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{as.sunburstDF}
\alias{as.sunburstDF}
\title{Prep data for a sunburst plot
Shamelessly copied from https://stackoverflow.com/a/58481176}
\usage{
as.sunburstDF(data, value_column = NULL, add_root = FALSE)
}
\arguments{
\item{data}{A dataframe/tibble}

\item{value_column}{The name of the column containing the value to be plotted}

\item{add_root}{Include a "Total" node in the center of the plot?}
}
\value{
A tibble properly formatted for use in a Plotly sunburst plot
}
\description{
Prep data for a sunburst plot
Shamelessly copied from https://stackoverflow.com/a/58481176
}
