% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{changeInCover_ggplot}
\alias{changeInCover_ggplot}
\title{Helper function for plotting cover change in ggplot}
\usage{
changeInCover_ggplot(data, max_lim)
}
\arguments{
\item{data}{Data to plot}

\item{max_lim}{maximum value in data}
}
\value{
ggplot object
}
\description{
Helper function for plotting cover change in ggplot
}
