% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ClearPACNVegCache}
\alias{ClearPACNVegCache}
\title{Clear cached data}
\usage{
ClearPACNVegCache(silent = FALSE)
}
\arguments{
\item{silent}{Silence feedback message?}
}
\value{
\code{TRUE} if cache was cleared, \code{FALSE} if no cache found
}
\description{
Clear cached data
}
