% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{DownloadAGOLAttachments}
\alias{DownloadAGOLAttachments}
\title{Download AGOL attachments (JPG or PNG)}
\usage{
DownloadAGOLAttachments(
  feature_layer_url,
  agol_username,
  agol_password = keyring::key_get("AGOL", agol_username),
  dest_folder = "photo_downloads",
  custom_name = TRUE,
  prefix = "photo",
  append_id = TRUE,
  sep = "_",
  join_cols = c(),
  after_date_filter = NA,
  only_staff = FALSE,
  test_run = FALSE
)
}
\arguments{
\item{feature_layer_url}{URL of the feature layer containing attachments (including layer ID). \emph{Do not} include a slash at the end.}

\item{agol_username}{AGOL username for a headless account with permissions to view the feature layer.}

\item{agol_password}{Do not hardcode passwords in your saved code! By default, assumes that the password was stored using \code{\link[keyring:key_get]{keyring::key_set()}} with \code{service = AGOL}.}

\item{dest_folder}{Folder in which to save downloaded attachments}

\item{custom_name}{Use prefix and sep to create filenames from attachment ID's? If \code{FALSE}, will use the filename as stored in AGOL. If you are not customizing your filenames when saving attachments, this may cause problems due to duplicate filenames.}

\item{prefix}{Prefix for photo file names. Can either be a character string (to set the same prefix for all files) or an unquoted column name from the attachment table.}

\item{append_id}{Append attachment id to custom prefix? Ignored if \code{custom_name == FALSE}}

\item{sep}{Separator character to use between prefix and photo ID in file names. Ignored if \code{custom_name} == FALSE}

\item{join_cols}{Named vector in the format \code{c("foreign key column to data table in attachment table" = "primary key column of data table")} where the attachment table is the table that stores attachments only (this is mostly hidden from view in AGOL) and the data table is the table for which attachments have been enabled. If you aren't sure what the foreign key column is called, leave this argument empty and set \code{test_run = TRUE}. The function will try to guess the join columns, but if it gets it wrong, it will return the attachment table without the accompanying data table. At that point, you can check the name(s) of the foreign key column(s) and re-run your code with the join columns specified. They should be something along the lines of "parentGlobalId" or "parentObjectId".}

\item{after_date_filter}{POSIXct date object. Only records after date will be returned.}

\item{only_staff}{if true, downloads just the staff photos.}

\item{test_run}{If \code{TRUE}, returns attachment data and proposed file locations without actually downloading and saving attachments.}
}
\value{
Tibble of attachment data, including paths to saved files.
}
\description{
From MOJN/Sarah Wright. Currently only works with jpeg or png images.
}
\examples{
\dontrun{
# Basic use
DownloadAGOLAttachments("https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_SV_OverviewPhotos/FeatureServer/0", agol_username = "mojn_data")

# Include data from table for which attachments are enabled
DownloadAGOLAttachments("https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_SV_OverviewPhotos/FeatureServer/0",
                        agol_username = "mojn_data",
                        join_cols = c("parentGlobalId" = "globalID"))
}
}
