% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIPS.R
\name{EIPS_cover_x_freq}
\alias{EIPS_cover_x_freq}
\title{Plot species cover x frequency for all species (facet by species)}
\usage{
EIPS_cover_x_freq(sample_frame, transect_number, save_folder)
}
\arguments{
\item{sample_frame}{EIPS sampling frame name}

\item{transect_number}{transect number (must be only one transect)}

\item{save_folder}{folder where .png plot will be saved}
}
\value{
graph of species found in transect across all cycles.
Optionally save individual .png file (if save_folder specified).
}
\description{
Plot species cover x frequency for all species (facet by species)
}
\details{
X axis is Cover Classes, y axis is frequency along transect
}
\examples{
\dontrun{
qc_EIPS_cover_x_freq(sample_frame =  "Mauna Loa", transect_number = 3, save_folder = "C:/Users/JJGross/OneDrive/Downloads")
}
}
