% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expect_dataframe_equal}
\alias{expect_dataframe_equal}
\title{Test for dataframe equivalence}
\usage{
expect_dataframe_equal(
  result,
  expected,
  ignore_col_order = FALSE,
  ignore_row_order = TRUE
)
}
\arguments{
\item{result}{Actual data frame}

\item{expected}{Expected data frame}

\item{ignore_col_order}{Ignore order of columns in dataframe? Defaults to FALSE.}

\item{ignore_row_order}{Ignore order of rows in dataframe? Defaults to TRUE.}
}
\value{
If test passes, nothing. If it fails, description of failure.
}
\description{
Test for dataframe equivalence
}
