% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{filename}
\alias{filename}
\title{Generate filename for icon generated by \code{\link[=pchIcons]{pchIcons()}}}
\usage{
filename(pch, col, bg)
}
\arguments{
\item{pch}{Plot character code (\href{http://www.sthda.com/english/wiki/r-plot-pch-symbols-the-different-point-shapes-available-in-r}{full list here})}

\item{col}{Color}

\item{bg}{Background color}
}
\value{
Filename and path
}
\description{
Helper function for \code{\link[=pchIcons]{pchIcons()}}
}
