% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FilterPACNVeg}
\alias{FilterPACNVeg}
\title{Filter PACN data}
\usage{
FilterPACNVeg(
  data_name,
  park,
  sample_frame,
  community,
  year,
  cycle,
  plot_type,
  is_qa_plot = FALSE,
  transect_type,
  species_code,
  sci_name,
  nativity,
  certified,
  verified,
  case_sensitive = FALSE,
  silent = FALSE
)
}
\arguments{
\item{data_name}{The name of the data table (see \code{names(GetColSpec())} for valid options)}

\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{year}{Monitoring year}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{is_qa_plot}{Whether the plots are QA plots or not (TRUE/FALSE)}

\item{transect_type}{Type of transect (fixed vs. rotational)}

\item{species_code}{6 letter species code}

\item{sci_name}{Scientific name}

\item{nativity}{Whether species are native (TRUE/FALSE)}

\item{certified}{Whether data are certified (TRUE/FALSE)}

\item{verified}{Whether data are verified (TRUE/FALSE)}

\item{case_sensitive}{Should non-numeric filters be treated as case-sensitive?}

\item{silent}{Suppress informational messages?}
}
\value{
A tibble (if data_name is provided) or a list of tibbles (if data_name is omitted)
}
\description{
Filter PACN data
}
\details{
All filters are optional. To ignore a filter, omit it or set it to NA.
}
\examples{
\dontrun{
all_hawaii_data <- FilterPACNVeg(park = c("KALA", "KAHO", "HAVO", "HALE"))
all_certified_data <- FilterPACNVeg(certified = TRUE)
native_kaho_understory <- FilterPACNVeg("Understory", park = "KAHO", nativity = "Native")
}
}
