% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{get_cycles}
\alias{get_cycles}
\title{Get cycles for a sample frame}
\usage{
get_cycles(sample_frame)
}
\arguments{
\item{sample_frame}{Name of sample frame}
}
\value{
A vector of cycles
}
\description{
Get cycles for a sample frame
}
