% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{get_years}
\alias{get_years}
\title{Get monitoring years for a sample frame}
\usage{
get_years(
  park,
  sample_frame,
  community,
  year,
  cycle,
  plot_type,
  plot_number,
  is_qa_plot,
  transect_type,
  species_code,
  sci_name,
  nativity,
  certified,
  verified,
  case_sensitive = FALSE,
  silent = FALSE
)
}
\arguments{
\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{year}{Monitoring year}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{is_qa_plot}{Whether the plots are QA plots or not (TRUE/FALSE) No argument defaults to FALSE. Can use "c(TRUE, FALSE)" without quotes to return both regular and QA plots.}

\item{transect_type}{Type of transect (fixed vs. rotational)}

\item{species_code}{6 letter species code}

\item{sci_name}{Scientific name}

\item{nativity}{Whether species are native (TRUE/FALSE)}

\item{certified}{Whether data are certified (TRUE/FALSE)}

\item{verified}{Whether data are verified (TRUE/FALSE)}

\item{case_sensitive}{Should non-numeric filters be treated as case-sensitive?}

\item{silent}{Suppress informational messages?}
}
\value{
A vector of years
}
\description{
If there are multiple monitoring years per cycle, assumes that the year with the most visits is the one that represents that cycle. You should verify that this is in fact the case for your data.
}
