% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{MapCoverTotal}
\alias{MapCoverTotal}
\title{Map Total Native vs. Non-native vegetation cover}
\usage{
MapCoverTotal(
  crosstalk = FALSE,
  crosstalk_group = "cover",
  combine_strata = TRUE,
  park,
  sample_frame,
  community,
  year,
  cycle = max(cycles),
  plot_type,
  is_qa_plot,
  certified,
  verified,
  silent = TRUE
)
}
\arguments{
\item{crosstalk}{If \code{TRUE}, returns a SharedData object for use with the \code{crosstalk} package. Otherwise, returns a normal tibble.}

\item{crosstalk_group}{If \code{crosstalk == TRUE}, the crosstalk group name to use. Ignored if \code{crosstalk == FALSE}.}

\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{is_qa_plot}{Whether the plots are QA plots or not (TRUE/FALSE) No argument defaults to FALSE. Can use "c(TRUE, FALSE)" without quotes to return both regular and QA plots.}

\item{certified}{Whether data are certified (TRUE/FALSE)}

\item{verified}{Whether data are verified (TRUE/FALSE)}
}
\value{
A leaflet map
}
\description{
Map Total Native vs. Non-native vegetation cover
}
\examples{
\dontrun{
MapCoverTotal(protocol = "FTPC") #currently only FTPC mapping supported
MapCoverTotal(park = "AMME")
MapCoverTotal(sample_frame = "Puu Alii")
}
}
