% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{MapPACNVeg}
\alias{MapPACNVeg}
\title{Map PACN vegetation plots and transects}
\usage{
MapPACNVeg(
  protocol = c("FTPC", "EIPS"),
  crosstalk = FALSE,
  crosstalk_group = "map",
  park,
  sample_frame,
  cycle,
  plot_type,
  is_qa_plot = FALSE,
  transect_type,
  certified,
  verified
)
}
\arguments{
\item{protocol}{Character vector indicating whether to include FTPC plots, EIPS transects, or both}

\item{crosstalk}{If \code{TRUE}, returns a SharedData object for use with the \code{crosstalk} package. Otherwise, returns a normal tibble.}

\item{crosstalk_group}{If \code{crosstalk == TRUE}, the crosstalk group name to use. Ignored if \code{crosstalk == FALSE}.}

\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{is_qa_plot}{Whether the plots are QA plots or not (TRUE/FALSE) No argument defaults to FALSE. Can use "c(TRUE, FALSE)" without quotes to return both regular and QA plots.}

\item{transect_type}{Type of transect (fixed vs. rotational)}

\item{certified}{Whether data are certified (TRUE/FALSE)}

\item{verified}{Whether data are verified (TRUE/FALSE)}
}
\value{
A leaflet map
}
\description{
Map PACN vegetation plots and transects
}
\examples{
\dontrun{
MapPACNVeg(protocol = "FTPC")
MapPACNVeg(park = "AMME")
}
}
