% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Presence.R
\name{master_spp_list}
\alias{master_spp_list}
\title{PACN Master Plant Species List}
\usage{
master_spp_list(
  db_path,
  park,
  presence_matrix = FALSE,
  sample_frame,
  community,
  year,
  cycle,
  plot_type,
  plot_number,
  nativity
)
}
\arguments{
\item{park}{Four letter unit code of park or can be "All"}

\item{presence_matrix}{stop function early to return sampling frame vs. species occurrence matrix}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{year}{Monitoring year}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{plot_number}{Plot number}

\item{nativity}{Whether species are native (TRUE/FALSE)}

\item{veg_species_db_path}{Full path to local copy of PACN_veg_species_list.accdb}
}
\value{
A tibble
}
\description{
Create a master species list with FTPC species occurrence by plot
Column "Field_ID" is also generated in table and can be used in Field Maps Plant Layer
}
\details{
park filter is required. All other filters are optional. To ignore a filter, omit it or set it to NA.
}
\examples{
\dontrun{
HALE_pacn_veg_spp_list <- master_spp_list(park = "HALE")
native_kaho_spp <- master_spp_list(park = "KAHO", nativity = "Native")
HAVO_spp_olaa_only <- master_spp_list(park = "HAVO", sample_frame = "Olaa")
}
}
