% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{pacn_dt}
\alias{pacn_dt}
\title{Create DT tables with consistent formatting}
\usage{
pacn_dt(data, record_count = TRUE, search = TRUE, ...)
}
\arguments{
\item{data}{A dataframe/tibble}

\item{record_count}{Display record count?}

\item{search}{Display search box?}

\item{...}{Additional arguments to pass to \link[DT:datatable]{DT::datatable}}
}
\description{
Create DT tables with consistent formatting
}
