% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacnvegetation-package.R
\docType{package}
\name{pacnvegetation-package}
\alias{pacnvegetation}
\alias{pacnvegetation-package}
\title{pacnvegetation: PACN Vegetation Data Analysis and Visualization}
\description{
This package is the home of all reusable code related to PACN vegetation protocols.
}
\author{
\strong{Maintainer}: Sarah Wright \email{sarah_wright@nps.gov}

Authors:
\itemize{
  \item Jacob Gross \email{jacob_gross@nps.gov}
}

}
\keyword{internal}
