% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{pchIcons}
\alias{pchIcons}
\title{Generate png icons for use in Leaflet}
\usage{
pchIcons(
  pch = 1,
  width = 30,
  height = 30,
  bg = "transparent",
  col = "black",
  lwd = 3,
  ...
)
}
\arguments{
\item{pch}{Plot character code (\href{http://www.sthda.com/english/wiki/r-plot-pch-symbols-the-different-point-shapes-available-in-r}{full list here})}

\item{width}{Width of character}

\item{height}{Height of character}

\item{bg}{Background color}

\item{col}{Color}

\item{...}{Additional arguments to \code{\link[graphics:points]{graphics::points()}}}
}
\value{
Filename(s) of icon(s)
}
\description{
Generate png icons for use in Leaflet
}
