% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photos.R
\name{process_photos}
\alias{process_photos}
\title{Watermark and process FTPC, EIPS, & Plant Photos}
\usage{
process_photos(
  AGOL_Layer,
  gdb_name,
  gdb_location,
  gdb_layer,
  return_first_table = FALSE,
  insert_first_table = NULL,
  add_watermark = FALSE,
  return_last_table = FALSE,
  insert_last_table = NULL
)
}
\arguments{
\item{AGOL_Layer}{"FTPC" "EIPS" or "Plants"}

\item{gdb_name}{Name of the geodatabase (Example: "EIPS_Olaa_Nahuku_20220323.gdb")}

\item{gdb_location}{File path to the geodatabase (Example: "C:/Users/JJGross/Documents/RData/PROJECTS/pacnvegetation/geodatabase")}

\item{gdb_layer}{The layer file inside the geodatabase (Example: "EIPS_Olaa_Nahuku_20220323")}

\item{return_first_table}{\code{default = FALSE} do not return table, process photos as normal; \code{return_table = TRUE} do not process photos, and return table instead.}

\item{insert_first_table}{\code{default = NULL} insert first table created in function. If R object supplied to \verb{insert_first_table = }, then user supplied table will be utilized in function instead of initial geodatabase.}

\item{add_watermark}{\code{default = FALSE} no watermark added}

\item{return_last_table}{\code{default = FALSE} do not return table, process photos as normal; \code{return_table = TRUE} do not process photos, and return table instead.}

\item{insert_last_table}{\code{default = NULL} insert final table created in function. If R object supplied to \verb{insert_last_table = }, then user supplied table will be used utilized in function instead of geodatabase.}
}
\value{
watermarked photos are saved to a folder "watermarked" inside the working directory. Use here() to determine current working directory.
}
\description{
Watermark and process FTPC, EIPS, & Plant Photos
}
\examples{
\dontrun{
process_photos(AGOL_Layer = "EIPS", gdb_name = "EIPS_Olaa_Nahuku_20220323.gdb", gdb_location = "C:/Users/JJGross/Documents/RData/PROJECTS/pacnvegetation/geodatabase", gdb_layer = "EIPS_Olaa_Nahuku_20220323")
}
}
