% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_presence.R
\name{qc_EIPS_spp_pres_dot_plot}
\alias{qc_EIPS_spp_pres_dot_plot}
\title{Plot all non-native species presence from EIPS transect}
\usage{
qc_EIPS_spp_pres_dot_plot(sample_frame, transect_number, save_folder)
}
\arguments{
\item{sample_frame}{EIPS sampling frame name}

\item{transect_number}{transect number (must be only one plot)}

\item{save_folder}{folder where .png plot will be saved}
}
\value{
graph of species found in transect across all cycles.
Optionally save individual .png file (if save_folder specified).
}
\description{
Plot all non-native species presence from EIPS transect
}
\details{
X axis is Cycle, y axis is species
}
\examples{
\dontrun{
qc_EIPS_spp_pres_dot_plot(sample_frame =  "Puu Alii", transect_number = 14, save_folder = "C:/Users/JJGross/OneDrive/Downloads")
}
}
