% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_presence.R
\name{qc_presence_complete}
\alias{qc_presence_complete}
\title{Check for completeness of species presence data by plot.}
\usage{
qc_presence_complete(
  all_records = TRUE,
  park,
  sample_frame,
  community,
  year,
  cycle,
  plot_type,
  is_qa_plot = FALSE,
  silent = TRUE
)
}
\arguments{
\item{all_records}{TRUE = Return all records (e.g. every cover hit), FALSE = Return 1st record of species per plot}

\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{year}{Monitoring year}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{is_qa_plot}{Whether the plots are QA plots or not (TRUE/FALSE)}

\item{silent}{Suppress informational messages?}
}
\value{
A tibble
}
\description{
'qc_presence_complete()' compares species recorded on the
Presence datasheet to the species recorded on other datasheets (Cover,
Small Trees, Shrub-belt, Large Trees). Species that may have been missed
on the Presence datasheet are listed.

All filters are optional. To ignore a filter, omit it or set it to NA.
}
\examples{
\dontrun{

species_missed <- qc_presence_complete()

HAVO_species_missed <- qc_presence_complete(park = "HAVO")

}
}
