% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_presence.R
\name{qc_sp_datasheets}
\alias{qc_sp_datasheets}
\title{Returns masterlist of all datasheets to determine where a species was recorded}
\usage{
qc_sp_datasheets(
  park,
  sample_frame,
  community,
  year,
  cycle,
  plot_type,
  plot_number,
  species_code,
  sci_name,
  nativity,
  is_qa_plot = FALSE,
  silent = TRUE
)
}
\arguments{
\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{year}{Monitoring year}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{species_code}{6 letter species code}

\item{sci_name}{Scientific name}

\item{nativity}{Whether species are native (TRUE/FALSE)}

\item{is_qa_plot}{Whether the plots are QA plots or not (TRUE/FALSE) No argument defaults to FALSE. Can use "c(TRUE, FALSE)" without quotes to return both regular and QA plots.}

\item{silent}{Suppress informational messages?}
}
\value{
A tibble
}
\description{
'qc_sp_datasheets()' returns all datasheets (Cover,
Small Trees, Shrub-belt, Large Trees) so that the location where a species
was recorded can be pinpointed.

All filters are optional. To ignore a filter, omit it or set it to NA.
}
\examples{
\dontrun{

sp_records_from_plot_46 <- qc_sp_datasheets(sample_frame = "Olaa", plot_number = 46, species_code = "MYRLES")


}
}
