% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_presence.R
\name{qc_spp_pres_dot_plot}
\alias{qc_spp_pres_dot_plot}
\title{Plot all species presence found within FTPC plot}
\usage{
qc_spp_pres_dot_plot(
  sample_frame,
  plot_number,
  save_folder,
  is_qa_plot = FALSE
)
}
\arguments{
\item{sample_frame}{FTPC sampling frame name}

\item{plot_number}{plot number (must be only one plot)}

\item{save_folder}{folder where .png plot will be saved}
}
\value{
individual .png files for each plot specified in sampling_frame + plot_number argument
}
\description{
Plot all species presence found within FTPC plot
}
\details{
X axis is Cycle, y axis is species
}
\examples{
\dontrun{
qc_spp_pres_dot_plot(sample_frame =  "Puu Alii", plot_number = 14, save_folder = "C:/Users/JJGross/OneDrive/Downloads")
}
}
