% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_spp_db}
\alias{read_spp_db}
\title{Read and return the species records from the PACN Veg Species Database
Vital Signs > 05_focal_terr_plant_communities > Data > Database > Veg_species_db}
\usage{
read_spp_db(db_paths)
}
\arguments{
\item{db_paths}{Database path (local path after downloading from sharepoint)}
}
\value{
A tibble
}
\description{
Read and return the species records from the PACN Veg Species Database
Vital Signs > 05_focal_terr_plant_communities > Data > Database > Veg_species_db
}
\examples{
\dontrun{
dbpath <- "C:/Users/JJGross/Documents/Databases_copied_local/Veg_species_db/PACN_veg_species_list_20230810.accdb"
pacnveg_master_spp_list <- read_spp_db(dbpath)

}
}
