% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_map_inventory.R
\name{read_vegmap_db}
\alias{read_vegmap_db}
\title{Combine species data from Veg Map Databases}
\usage{
read_vegmap_db(db_paths)
}
\arguments{
\item{db_paths}{Database path or list of database paths pointing to}
}
\value{
A tibble with one row per location and columns Protocol, Unit_Code, Sampling_Frame, Sample_Unit_Number, Lat, Long, Cycles, Years
}
\description{
Combine species data from Veg Map Databases
}
\examples{
\dontrun{
all_locs <- PlotAndTransectLocations()
eips_transects <- PlotAndTransectLocations("EIPS")
}
}
