% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReadCSV}
\alias{ReadCSV}
\title{Read data from a folder of csv files}
\usage{
ReadCSV(data_path)
}
\arguments{
\item{data_path}{A path to a folder containing the data in csv format}
}
\value{
A list of tibbles
}
\description{
Read data from a folder of csv files
}
