% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReadEIPS}
\alias{ReadEIPS}
\title{Read data from EIPS database(s)}
\usage{
ReadEIPS(db_paths)
}
\arguments{
\item{db_paths}{A path or vector of paths to Access databases containing EIPS data}
}
\value{
A list of tibbles
}
\description{
Read data from EIPS database(s)
}
