% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReadFTPC}
\alias{ReadFTPC}
\title{Read data from FTPC database}
\usage{
ReadFTPC(conn, TE_Species)
}
\arguments{
\item{conn}{Database connection object returned by \code{DBI::dbConnect}}

\item{TE_Species}{Include T&E species names (TRUE/FALSE).}
}
\value{
A list of tibbles
}
\description{
Read data from FTPC database
}
