% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RemoveSingleVisits}
\alias{RemoveSingleVisits}
\title{Remove data from plots with no revisits}
\usage{
RemoveSingleVisits(data)
}
\arguments{
\item{data}{A tibble/dataframe with columns Unit_Code, Sampling_Frame, Plot_Number, and Cycle}
}
\value{
The input data with  single-visit data removed
}
\description{
Remove data from plots with no revisits
}
\examples{
\dontrun{
data <- FilterPACNVeg("Understory")
data_with_revisits <- RemoveSingleVisits(data)
}
}
