% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{summarize_understory}
\alias{summarize_understory}
\title{Summarize Understory Cover Data by Nativity, Life_Form, or Species}
\usage{
summarize_understory(
  combine_strata = FALSE,
  plant_grouping,
  paired_change = FALSE,
  park,
  sample_frame,
  community,
  year,
  cycle,
  plot_type,
  plot_number,
  species_code,
  silent = FALSE
)
}
\arguments{
\item{combine_strata}{If \code{TRUE}, combine high and low strata into one stratum.}

\item{plant_grouping}{"None" = No Groups, all vegetation (native & non-native) is lumped together,
"Nativity" = Cover values lumped by native and non-native vegetation.
"Life_Form" = Cover values grouped into native and non-native lifeforms (e.g. Trees, Shrubs, Ferns).
"Species" = Cover values for individual plant species.}

\item{paired_change}{If \code{TRUE}, calculate change in percent cover. Note that this drops "Rotational" plots and uses "Fixed" plots only.}

\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{year}{Monitoring year}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{species_code}{6 letter species code}

\item{silent}{Suppress informational messages?}
}
\value{
Summary table of total Percent Cover by grouping or change in total Percent Cover by grouping.
}
\description{
Summarize Understory Cover Data by Nativity, Life_Form, or Species
}
\examples{
\dontrun{
native_and_non-native_cover <- summarize_understory(plant_grouping = "Nativity")
species_cover_change <- summarize_understory(plant_grouping = "Species", paired_change = TRUE)

}
}
