% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{totalCover_plotly}
\alias{totalCover_plotly}
\title{Helper function for plotting total cover in plotly}
\usage{
totalCover_plotly(data, max_lim)
}
\arguments{
\item{data}{Data to plot}

\item{max_lim}{maximum value in data}
}
\value{
html widget
}
\description{
Helper function for plotting total cover in plotly
}
