% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{UnderCombineStrata}
\alias{UnderCombineStrata}
\title{Combine Strata low & high}
\usage{
UnderCombineStrata(data)
}
\arguments{
\item{data}{An Understory tibble/dataframe with Species Point Data}
}
\value{
The Understory tibble/dataframe with one stratum "combined"
}
\description{
Combine Strata low & high
}
\examples{
\dontrun{
data <- FilterPACNVeg("Understory")
understory_with_one_stratum <- CombineStrata(data)
}
}
