% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{UnderNativityCover.plot.nat_v_non}
\alias{UnderNativityCover.plot.nat_v_non}
\title{Native V. Non-native Cover Plot for Understory data}
\usage{
UnderNativityCover.plot.nat_v_non(data, sample_cycle)
}
\arguments{
\item{data}{An Understory tibble/dataframe with Native and Non-native Cover}
}
\value{
Nat-v-Non Plot comparing Native and Non-native Cover (or Change in Cover)
}
\description{
Native V. Non-native Cover Plot for Understory data
}
\examples{
\dontrun{

Native_v_Nonnative_Plot <- UnderNativityCover.plot.nat_v_non(UnderNativityCover())
}
}
