% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{UnderNativityCover}
\alias{UnderNativityCover}
\title{Calculate Total Native & Nonnative Cover in Understory}
\usage{
UnderNativityCover(
  combine_strata = FALSE,
  paired_change = FALSE,
  rm_unk_natv = TRUE,
  crosstalk = FALSE,
  crosstalk_group = "cover",
  park,
  sample_frame,
  community,
  year,
  cycle,
  plot_type,
  paired_cycle = 1,
  silent = FALSE
)
}
\arguments{
\item{combine_strata}{If \code{TRUE}, don't split data into high and low strata. Otherwise, keep as is.}

\item{paired_change}{If \code{TRUE}, calculate change in percent cover across two cycles. If \code{cycle} is not specified, it will default to the last cycle.}

\item{rm_unk_natv}{Remove}

\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{year}{Monitoring year}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{paired_cycle}{Only required if \code{paired_change == TRUE}. The cycle number to compare to when calculating paired change.}

\item{silent}{Suppress informational messages?}
}
\value{
Summary table of total Percent Cover by Nativity
}
\description{
Calculate Total Native & Nonnative Cover in Understory
}
\examples{
\dontrun{
data <- FilterPACNVeg("Understory")
Native_Cover_Summary_table <- UnderNativityCoverTotal(data)
}
}
