% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{understory_cover_boxplot}
\alias{understory_cover_boxplot}
\title{Create boxplots of understory cover for all Sampling Frames in Dataset}
\usage{
understory_cover_boxplot(
  df,
  y = "Chg_Prior",
  nativity = "Native",
  cycle = 2,
  no_y_labs = FALSE
)
}
\arguments{
\item{df}{Dataframe from summarize_understory()}

\item{y}{y-variable from dataframe used for boxplot}

\item{nativity}{default = "Native"; can be changed to "Non-Native"}

\item{cycle}{default = 2; select one cycle, can be 2 or 3}

\item{no_y_labs}{default = FALSE; if true, removes y-axis tick marks and labels}
}
\value{
ggplot boxplot
}
\description{
Create boxplots of understory cover for all Sampling Frames in Dataset
}
\examples{
\dontrun{
ncyc3 <- understory_cover_boxplot(nat_chg_und, nativity = "Non-Native", cycle = 3, no_y_labs = TRUE)

}
}
