% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{understoryBarCover}
\alias{understoryBarCover}
\title{Bar plot of avg. \% cover vs cycle}
\usage{
understoryBarCover(
  sample_frame,
  cycle,
  crosstalk_filters = TRUE,
  colors = c("#F8573A", "#F4C47B", "#28468B", "#AED5CB")
)
}
\arguments{
\item{crosstalk_filters}{Include dropdowns to filter on species, nativity, and mgmt unit?}

\item{colors}{Character vector of color names or hex values. This can be a named vector with names corresponding to each management unit, or you can provide an unnamed vector of any number of colors to create a gradient.}
}
\value{
An HTML object (if \code{crosstalk_filters == TRUE}) or a plotly object.
}
\description{
Broken down by management unit
}
