% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{understorySpeciesCover}
\alias{understorySpeciesCover}
\title{Title}
\usage{
understorySpeciesCover(
  sample_frame,
  cycle,
  group_by = c("GROUP_COL", "Cycle", "Nativity")
)
}
\arguments{
\item{group_by}{Columns to group by. Must be a subset of \code{c("GROUP_COL", "Cycle", "Nativity")}}
}
\value{
A tibble
}
\description{
Title
}
