% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{understorySpeciesCover2}
\alias{understorySpeciesCover2}
\title{Title}
\usage{
understorySpeciesCover2(sample_frame, cycle, group_by = c("Zone", "Cycle"))
}
\arguments{
\item{group_by}{Columns to group by. Must be a subset of \code{c("GROUP_COL", "Cycle")}}
}
\value{
A tibble
}
\description{
Title
}
\examples{
\dontrun{

Compare understorySpeciesCover2() vs. summarize_understory()

comp_usc2 <- understorySpeciesCover2(sample_frame = "Olaa", cycle = 3)

comp_su <- summarize_understory(plant_grouping = "Species", sample_frame = "Olaa", cycle = 3, combine_strata = TRUE)

comp_su_summarize <- comp_su |>
 group_by(Unit_Code, Sampling_Frame, Nativity, Life_Form, Scientific_Name, Code) |>
 dplyr::summarize(n = dplyr::n(),
                  plots_present = sum(Cover > 0),
                  Avg_Cover = round(mean(Cover), 3),
                  Std_Dev = round(sd(Cover), 3),
                  .groups = "drop")

}
}
