% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{understorySunburst}
\alias{understorySunburst}
\title{Sunburst plot of understory percent cover}
\usage{
understorySunburst(sample_frame, cycle, mgmt_unit = TRUE, colors = "default")
}
\arguments{
\item{sample_frame}{Name of sample frame}

\item{cycle}{Monitoring cycle}

\item{mgmt_unit}{Include management unit?}

\item{colors}{Character vector of color names or hex values. This can be a named vector with names corresponding to the values in the center of the plot (the first column specified in \code{group_by}), or you can provide an unnamed vector of any number of colors to create a gradient.}
}
\value{
A plotly object
}
\description{
Sunburst plot of understory percent cover
}
