% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{v_cover_bar_spp_plot}
\alias{v_cover_bar_spp_plot}
\title{Bar plot of species \% cover for each sampling plot}
\usage{
v_cover_bar_spp_plot(sample_frame, crosstalk_filters = TRUE, crosstalk_group)
}
\arguments{
\item{sample_frame}{(required)}

\item{crosstalk_filters}{Include dropdowns to filter on species, nativity, and mgmt unit?}

\item{crosstalk_group}{Help distinguish crosstalk iterations}
}
\value{
An HTML object (if \code{crosstalk_filters == TRUE}) or a plotly object.
}
\description{
Can be placed side-by-side to compare two plots
}
