% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{v_cover_bar_stats}
\alias{v_cover_bar_stats}
\title{Stats Bar Plot for Understory Cover by Species, Life form, Nativity, or no grouping (all cover lumped)}
\usage{
v_cover_bar_stats(
  combine_strata = FALSE,
  plant_grouping = "Nativity",
  species_filter,
  paired_change = FALSE,
  measurement = "Cover",
  park,
  sample_frame,
  community,
  year,
  cycle,
  plot_type,
  plot_number,
  filter_Code,
  silent = FALSE,
  return_n = FALSE,
  remove_unknown = TRUE
)
}
\arguments{
\item{combine_strata}{If \code{TRUE}, combine high and low strata into one stratum.}

\item{plant_grouping}{"None" = No Groups, all vegetation (native & non-native) is lumped together,
"Nativity" = Cover values lumped by native and non-native vegetation.
"Life_Form" = Cover values grouped into native and non-native lifeforms (e.g. Trees, Shrubs, Ferns).
"Species" = Cover values for individual plant species.}

\item{paired_change}{If \code{TRUE}, calculate change in percent cover. Note that this drops "Rotational" plots and uses "Fixed" plots only.}

\item{measurement}{"Cover" = Total \% Cover for plant_grouping.
"Chg_Prior = Difference between Cover for a selected cycle and the previous cycle.
"Years_Prior" = Number of years between sampling cycles.
"Chg_Per_Year" = Chg_Prior divided by Years_Prior.}

\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{year}{Monitoring year}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{silent}{Suppress informational messages?}

\item{remove_unknown}{"default == TRUE. Removes any Life_form with value == "Unknown".
Also removes the non-vegetated points from final display.}
}
\value{
graph (gglot) of total Percent Cover by Nativity
}
\description{
Stats Bar Plot for Understory Cover by Species, Life form, Nativity, or no grouping (all cover lumped)
}
\examples{
\dontrun{
subalpine_mean_nativity_cover <- v_cover_bar_stats(community = "Subalpine Shrubland")

}
}
