% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{v_cover_plot_bar_nativity}
\alias{v_cover_plot_bar_nativity}
\title{Stats Bar Plot for Understory Cover by Nativity}
\usage{
v_cover_plot_bar_nativity(
  combine_strata = FALSE,
  paired_change = FALSE,
  park,
  sample_frame,
  community,
  year,
  cycle,
  plot_type,
  plot_number,
  filter_Code,
  silent = FALSE
)
}
\arguments{
\item{combine_strata}{If \code{TRUE}, combine high and low strata into one stratum.}

\item{paired_change}{If \code{TRUE}, calculate change in percent cover. Note that this drops "Rotational" plots and uses "Fixed" plots only.}

\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{year}{Monitoring year}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{silent}{Suppress informational messages?}
}
\value{
graph (gglot) of total Percent Cover by Nativity
}
\description{
Stats Bar Plot for Understory Cover by Nativity
}
\examples{
\dontrun{
subalpine_mean_nativity_cover <- v_cover_plot_bar_nativity(community = "Subalpine Shrubland")

}
}
