% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIPS.R
\name{v_EIPS_cover_class2percent}
\alias{v_EIPS_cover_class2percent}
\title{Convert EIPS Cover Class column to Range of Cover or vice-versa}
\usage{
v_EIPS_cover_class2percent(df, cover_column)
}
\value{
Filename and path
}
\description{
Input a column with values 0-7 for Braun-Blanquet cover classes and
returns two percent cover columns (min and max percent cover). Note that
Cover Class "0" returns 0 for min and max, while cover class 7 will return 0.75
for min and 1.0 for max.
}
\examples{
\dontrun{
MapPACNVeg(protocol = "FTPC")
MapPACNVeg(park = "AMME")
}
}
