% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIPS.R
\name{v_EIPS_map_interstation}
\alias{v_EIPS_map_interstation}
\title{Map EIPS data}
\usage{
v_EIPS_map_interstation(
  .data,
  parameter,
  change = FALSE,
  agol_sample_frame = "*"
)
}
\value{
leaflet map with crosstalk selection
}
\description{
Input a dataframe, metric, and a column to be mapped, returns a map using
the column specified to color the metric.
}
\examples{
\dontrun{
v_EIPS_map()
v_EIPS_map()
}
}
