% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIPS.R
\name{v_EIPS_prep}
\alias{v_EIPS_prep}
\title{Prep and check EIPS data and add helpful columns}
\usage{
v_EIPS_prep(
  park,
  sample_frame,
  community,
  year,
  cycle,
  transect_type,
  verified,
  certified
)
}
\value{
dataframe
#' @examples
#' \dontrun{
v_EIPS_prep(sample_frame = "Olaa")
v_EIPS_prep(park = "AMME")
}
}
\description{
Pulls EIPS dataset from database cache and performs some
basic data preparation.
}
