% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Presence.R
\name{v_presence_table}
\alias{v_presence_table}
\title{FTPC Species Presence Table}
\usage{
v_presence_table(sample_frame, cycle, table_type = "html")
}
\arguments{
\item{sample_frame}{Name of sampling frame}

\item{table_type}{default is "html" (other option is "tibble")}
}
\value{
A html table
}
\description{
Species presence per plot per year for all years available within a sampling frame.
min and max plot percentage shown with sparkline showing presence trend for all years
}
\examples{
\dontrun{
FTPC_Olaa_Presence <- v_presence_table(sample_frame, table_type = "html")
}
}
