% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photos.R
\name{watermark}
\alias{watermark}
\title{Watermark and process FTPC, EIPS, & Plant Photos}
\usage{
watermark(x, new_folder)
}
\arguments{
\item{x}{GIS table}

\item{new_folder}{folder where watermarked photos are saved}
}
\value{
an exported watermarked photo
}
\description{
Watermark and process FTPC, EIPS, & Plant Photos
}
\examples{
\dontrun{
watermark(GIS_Table, "watermarked")
}
from the Exif data.
}
