% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photos.R
\name{watermark_no}
\alias{watermark_no}
\title{Process FTPC, EIPS, & Plant Photos without a watermark}
\usage{
watermark_no(x, new_folder)
}
\arguments{
\item{x}{GIS table}

\item{new_folder}{folder where watermarked photos are saved}
}
\value{
an exported watermarked photo
}
\description{
Process FTPC, EIPS, & Plant Photos without a watermark
}
\examples{
\dontrun{
watermark_no(GIS_Table, "new_folder")
}
from the Exif data.
}
