% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{get_years}
\alias{get_years}
\title{Get monitoring years for a sample frame}
\usage{
get_years(sample_frame)
}
\arguments{
\item{sample_frame}{Name of sample frame}
}
\value{
A vector of years
}
\description{
If there are multiple monitoring years per cycle, assumes that the year with the most visits is the one that represents that cycle. You should verify that this is in fact the case for your data.
}
