% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photos.R
\name{process_photos}
\alias{process_photos}
\title{Watermark and process FTPC, EIPS, & Plant Photos}
\usage{
process_photos(
  AGOL_Layer,
  gdb_name,
  gdb_location,
  gdb_layer,
  return_table = FALSE
)
}
\arguments{
\item{AGOL_Layer}{"FTPC" "EIPS" or "Plants"}

\item{gdb_name}{Name of the geodatabase (Example: "EIPS_Olaa_Nahuku_20220323.gdb")}

\item{gdb_location}{File path to the geodatabase (Example: "C:/Users/JJGross/Documents/RData/PROJECTS/pacnvegetation/geodatabase")}

\item{gdb_layer}{The layer file inside the geodatabase (Example: "EIPS_Olaa_Nahuku_20220323")}

\item{return_table}{FALSE = process photos; TRUE = do not process photos, and return table instead}
}
\value{
watermarked photos are saved to a folder "watermarked" inside the working directory. Use here() to determine current working directory.
}
\description{
Watermark and process FTPC, EIPS, & Plant Photos
}
\examples{
\dontrun{
process_photos(AGOL_Layer = "EIPS", gdb_name = "EIPS_Olaa_Nahuku_20220323.gdb", gdb_location = "C:/Users/JJGross/Documents/RData/PROJECTS/pacnvegetation/geodatabase", gdb_layer = "EIPS_Olaa_Nahuku_20220323")
}
}
