% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_map_inventory.R
\name{read_vegmap_db}
\alias{read_vegmap_db}
\title{Combine species data from Veg Map Databases}
\usage{
read_vegmap_db(db_paths)
}
\arguments{
\item{db_paths}{Database path or list of database paths pointing to}
}
\value{
A tibble with one row per location and columns Protocol, Unit_Code, Sampling_Frame, Sample_Unit_Number, Lat, Long, Cycles, Years
}
\description{
Combine species data from Veg Map Databases
}
\examples{
\dontrun{

hi_vegmap_db_paths <- c("C:/Users/JJGross/OneDrive - DOI/Documents/Veg_Map_Data/havodata.accdb",
                       "C:/Users/JJGross/OneDrive - DOI/Documents/Veg_Map_Data/haledata.accdb",
                       "C:/Users/JJGross/OneDrive - DOI/Documents/Veg_Map_Data/kahodata.mdb",
                       "C:/Users/JJGross/OneDrive - DOI/Documents/Veg_Map_Data/kaladata.mdb",
                       "C:/Users/JJGross/OneDrive - DOI/Documents/Veg_Map_Data/puhedata.mdb",
                       "C:/Users/JJGross/OneDrive - DOI/Documents/Veg_Map_Data/puhodata.mdb")

# Veg map & species locations ----

Hawaii_vegmap_data <- read_vegmap_db(hi_vegmap_db_paths)
}
}
