% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{UnderNativityCover.plot.nat_v_non}
\alias{UnderNativityCover.plot.nat_v_non}
\title{Native V. Non-native Cover Plot for Understory data}
\usage{
UnderNativityCover.plot.nat_v_non(
  combine_strata = FALSE,
  paired_change = FALSE,
  sample_cycle,
  park,
  sample_frame,
  community,
  year,
  cycle,
  plot_type,
  silent = FALSE
)
}
\arguments{
\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{silent}{Suppress informational messages?}
}
\value{
Nat-v-Non Plot comparing Native and Non-native Cover (or Change in Cover)
}
\description{
Native V. Non-native Cover Plot for Understory data
}
\examples{
\dontrun{

Native_v_Nonnative_Plot <- UnderNativityCover.plot.nat_v_non()
}
}
