% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{UnderNativityCover}
\alias{UnderNativityCover}
\title{Calculate Total Native & Nonnative Cover in Understory}
\usage{
UnderNativityCover(
  combine_strata = FALSE,
  paired_change = FALSE,
  park,
  sample_frame,
  community,
  year,
  cycle,
  plot_type,
  silent = FALSE
)
}
\arguments{
\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{year}{Monitoring year}

\item{cycle}{Monitoring cycle}

\item{plot_type}{Type of plot (fixed vs. rotational)}

\item{silent}{Suppress informational messages?}
}
\value{
Summary table of total Percent Cover by Nativity
}
\description{
Calculate Total Native & Nonnative Cover in Understory
}
\examples{
\dontrun{
data <- FilterPACNVeg("Understory")
Native_Cover_Summary_table <- UnderNativityCoverTotal(data)
}
*Dead Still present in Understory Data*
}
