% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_understory.R
\name{understorySpeciesCover2}
\alias{understorySpeciesCover2}
\title{Title}
\usage{
understorySpeciesCover2(sample_frame, cycle, group_by = c("Zone", "Cycle"))
}
\arguments{
\item{group_by}{Columns to group by. Must be a subset of \code{c("GROUP_COL", "Cycle")}}
}
\value{
A tibble
}
\description{
Title
}
