% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WritePACNVeg}
\alias{WritePACNVeg}
\title{Save PACN vegetation data as a set of .csv files}
\usage{
WritePACNVeg(
  dest.folder,
  create.folders = FALSE,
  overwrite = FALSE,
  park,
  sample_frame,
  community,
  certified,
  verified
)
}
\arguments{
\item{dest.folder}{The folder in which to save the .csv files.}

\item{create.folders}{Should \code{dest.folder} be created automatically if it doesn't exist? Defaults to \code{FALSE}.}

\item{overwrite}{Should existing data be automatically overwritten? Defaults to \code{FALSE}.}

\item{park}{Four letter unit code of park(s)}

\item{sample_frame}{Name of sample frame}

\item{community}{Name of plant community}

\item{certified}{Whether data are certified (TRUE/FALSE)}

\item{verified}{Whether data are verified (TRUE/FALSE)}
}
\value{
None.
}
\description{
Save PACN vegetation data as a set of .csv files
}
\examples{
\dontrun{
LoadPACNVeg("pacnveg", "path/to/access.mdb")
WritePACNVeg("folder/for/csv/data", create.folders = TRUE)
}
}
